Rename copyslicemap to copyboundary
