+++ analyzers/copyboundary/copy_boundary_test.go
@@ -1,4 +1,4 @@
-package copyslicemap
+package copyboundary
 
 import (
 	"testing"
