+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/gophersbd/gouberfmt/analyzers/copyslicemap"
+	"github.com/gophersbd/gouberfmt/analyzers/copyboundary"
 	"github.com/gophersbd/gouberfmt/analyzers/interfacepointer"
 	"github.com/gophersbd/gouberfmt/analyzers/mutexpointer"
 
@@ -12,6 +12,6 @@ func main() {
 	multichecker.Main(
 		interfacepointer.Analyzer,
 		mutexpointer.Analyzer,
-		copyslicemap.Analyzer,
+		copyboundary.Analyzer,
 	)
 }
